moves headers in the detail space to public space
Closes #5

This commit is responsible for moving all public-facing names outside of
the detail header space and into the public space for individual
inclusion. This makes it possible to include a single component of the
library, rather than needing to include a header that includes
everything else also. Anything that needed to be cleaned up as a result
(e.g. include paths, header guards) have also been adjusted.

This commit introduces a few new headers, which assume the path of the
just-moved-headers. These headers are designed to contain as many
details as possible until modules arrive.