enables clang-tidy, massive clean-up
Enabling clang-tidy forced a huge clean-up to the codebase, which is
actioned simultaneously. clang-tidy doesn't get along with doctest, so
the project has (hopefully temporarily) moved to range-v3's simple_test
test-framework.

clang-tidy is also unable to handle the implementation of contracts that
was present, due to the way that it handles unambiguously throwing in
noexcept-specified functions. As such, Contracts Consolation was
designed, exported to a thrid-party library, and imported into this
project for usage.

Cleaning up function names was also actioned along the way (e.g. from
`models_Invocable` to `models_invocable`, etc.).