tidies config, prepares for MSVC
The build system ecosystem was in need of a serious clean-up; lots of
unused CMake was removed or tuned specifically for this project.

The project should also start to turn its attention to MSVC, since that
compiler is starting to support concepts. This commit readies some, but
not all, of that work (e.g. the MSVC Conan profile now works with MSVC
16.3 Preview 4).