adds a check to ensure constexpr-checking evaluates Boolean expressions
CJDB_CONSTEXPR_CHECK was able to evaluate expressions that are
convertible to `bool`, but this also includes partially-formed
expressions (such as `CJDB_CONSTEXPR_CHECK(x)` when `x` is an `int`),
lambdas, and function pointers.

Since the intention of passing an invocable to `CJDB_CONSTEXPR_CHECK`
is probably for it to be called, this check has been put in place to
flag any expressions where the author has passed in a named lambda, but
forgotten to actually call it.