revises the CI so it depends more on Cirrus CI infra (#18)
revises the CI so it depends more on Cirrus CI infra

The existing infrastructure depends on building clang-concepts and GCC
(order dependent), which is slow and cumbersome. This commit changes the
behaviour so that clang-concepts and GCC can be built independently from
one another, and so that Cirrus will always build a Docker image based
on the latest built compilers.