adds existence tests for [iterators] and [ranges]
Each header should have a set of tests to ensure that the
publicly-facing names are, in fact, available for use. These are called
'existence' tests, as they check for the existence of certain names.

As some names weren't publicly available in the corresponding standard
headers, those names have also been added.